requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.3.1
Werkzeug==2.0.1
pymongo==3.12.0
bcrypt==3.2.0
nltk==3.6.2
Pillow==9.0.1
pytesseract==0.3.7
